remove validações de papéis desnecessários e move arquivo
