+++ products-ms/src/main/kotlin/puc/domain/users/services/UserService.kt
@@ -2,19 +2,13 @@ package puc.domain.users.services
 
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
-import puc.application.consumers.IUserClient
+import puc.application._shared.IUserClient
 import puc.domain.users.model.User
 import kotlin.jvm.optionals.getOrNull
 
 @Service
 class UserService(val userClient: IUserClient): IUserService {
     val logger = LoggerFactory.getLogger(this.javaClass)!!
-
-    /* val userClient: UserClient = mock<UserClient>().apply {
-        val user = Optional.of(User(1, "JohnDoe", setOf("ROLE_USER")))
-        whenever(this.findById(1)).thenReturn(user);
-        whenever(this.getMe()).thenReturn(user);
-    } */
 
 
     override fun getAuthenticatedUser(): User? {
