+++ products-ms/src/main/resources/application.properties
@@ -12,6 +12,6 @@ spring.redis.port=6379
 client.user.url=http://localhost:8083
 client.user.path=/users
 
-logging.level.puc.application.consumers.IUserClient=DEBUG
+logging.level.puc.application._shared.IUserClient=DEBUG
 spring.cloud.openfeign.client.config.user-client.dismiss404=true
 spring.cloud.openfeign.client.config.user-client.logger-level=basic
